:recycle: refactor(api): move session summaries usecase to src
