+++ api/src/certification/session-management/domain/usecases/find-paginated-certification-center-session-summaries.js
@@ -1,4 +1,4 @@
-import { ForbiddenAccess } from '../../../src/shared/domain/errors.js';
+import { ForbiddenAccess } from '../../../../../src/shared/domain/errors.js';
 
 const findPaginatedCertificationCenterSessionSummaries = async function ({
   userId,
