+++ api/tests/certification/session-management/unit/application/certification-centers-session-summaries-controller.test.js
@@ -1,5 +1,5 @@
 import { certificationCenterController } from '../../../../../src/certification/session-management/application/certification-centers-session-summaries-controller.js';
-import { usecases } from '../../../../../lib/domain/usecases/index.js';
+import { usecases } from '../../../../../src/certification/session-management/domain/usecases/index.js';
 import { domainBuilder, expect, hFake, sinon } from '../../../../test-helper.js';
 
 describe('Unit | Controller | certifications-center-controller', function () {
